% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{query_all}
\alias{query_all}
\title{Title: Query for word matches to query words}
\usage{
query_all(x, y)
}
\arguments{
\item{x}{Character vector to query}

\item{y}{List of query words}
}
\value{
A character vector of class character
}
\description{
The function queries a given text for words with either an
exact match, a prefix or suffix of characters specified in a list of query words.
}
\examples{
txt <- c("I", "am", "so", "happy", "that", "you" "are", "using", "this", "function")
query_words <- c("am", "us*", "that", "*tion")

query_all(txt, query_words)

}
