% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{query_equal}
\alias{query_equal}
\title{Title: Query for word matches}
\usage{
query_equal(x, y)
}
\arguments{
\item{x}{Character vector to query}

\item{y}{List of query words}
}
\value{
A character vector of class character
}
\description{
The function queries a given text for characters with an
exact match to those specified in a list of query words
}
\examples{
txt <- c("I", "am", "so", "happy")
query_words <- "am"

query_equal(txt, query_words)

}
