% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{query_prefix}
\alias{query_prefix}
\title{Title: Query for word prefix}
\usage{
query_prefix(x, y)
}
\arguments{
\item{x}{Character vector to query}

\item{y}{List of words with prefix annotation "word*"}
}
\value{
A character vector of class character
}
\description{
The function queries a given text for words begining with a specified prefix.
}
\examples{
txt <- c("I", "am", "so", "happy")
prefix_words <- "happ*"

query_prefix(txt, pos_words)

}
