% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{query_suffix}
\alias{query_suffix}
\title{Title: Query for word suffix}
\usage{
query_suffix(x, y)
}
\arguments{
\item{x}{Character vector to query}

\item{y}{List of words with suffix annotation "*word"}
}
\value{
A character vector of class character
}
\description{
The function queries a given text for words ending with a specified suffix.
}
\examples{
txt <- c("She", "wouldn't", "be", "happy")
suffix_words <- c("*py", "*'t")

query_suffix(txt, suffix_words)

}
