% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{split_sequential}
\alias{split_sequential}
\title{Title: Sequential division of words into single elements}
\usage{
split_sequential(x)
}
\arguments{
\item{x}{A txt.file or string}
}
\value{
character vector with separated words
}
\description{
The function can process a text file or string. It performs sequential division
of words into single elements, splitting is based on typical character separators
including; tab, space, comma, semicolon and full stop.
}
\examples{
txt <- c("I am so happy")

txt <- split_sequential(txt)

}
