% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{txt_low_split}
\alias{txt_low_split}
\title{Title: Sequential word division and conversion to lower case letters}
\usage{
txt_low_split(x)
}
\arguments{
\item{x}{A txt.file or string}
}
\value{
character vector with separated words and lowercase letters
}
\description{
The function can process a text file or string. It performs sequential division
of words into single elements, splitting is based on typical character separators
including; tab, space, comma, semicolon and full stop. The output vector has all lower case letters.
}
\examples{
txt <- c("I am so happy")

txt_low_split(txt)

}
