% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{wordcount}
\alias{wordcount}
\title{Title: Word counts of query words}
\usage{
wordcount(x, y)
}
\arguments{
\item{x}{Character vector (to query)}

\item{y}{List of query words}
}
\value{
counts for each query word
}
\description{
The function provides word count (the number of times the word appears)
in a given text for a specified list of query words.
}
\examples{
txt <- c("I", "am", "so", "happy")
pos_words <- "happy"

wordcount(txt, pos_words)

}
