% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{wordratio}
\alias{wordratio}
\title{Title: Ratio of the sum of query words}
\usage{
wordratio(x, y, z)
}
\arguments{
\item{x}{Character vector (to query)}

\item{y}{List of query words}

\item{z}{List of query words}
}
\value{
numerical ratio of wordsum(y)/worsum(z)
}
\description{
The function provides a numeric ratio of the sum of total
query words of two separate query vectors.
}
\examples{
txt <- c("I", "am", "so", "proud", "and", "scared")
pos_words <- "proud"
neg_words <- "scared"

wordratio(txt, pos_words, neg_words)

}
