% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFUN.R
\name{wordsum}
\alias{wordsum}
\title{Title: Sum of query words}
\usage{
wordsum(x, y)
}
\arguments{
\item{x}{Character vector (to query)}

\item{y}{List of query words}
}
\value{
numerical sum
}
\description{
The function provides the sum of wordcounts.
}
\examples{
txt <- c("I", "am", "so", "happy")
pos_words <- "happy"

wordsum(txt, pos_words)

}
